feat: cleaned + removed window calls
